7mo commit: Optimizaciones y BD de usuarios y comisiones agregada

+++ src/models/product.model.js
@@ -1,42 +1,3 @@
-/* ESTRUCTURA DE DATOS DEL PRODUCTO
-El producto debe tener la siguiente estructura:
-    {
-        "id": "8GcQtrHKoCKb5UL45sni", // ID único del producto
-        "store": "pies a la moda", // Nombre de la tienda
-        "category": "calzado", // Categoría del producto
-        "commision": 5000, // Comisión por venta
-        "article": "campus importadas", // Nombre del artículo
-        "price": 30000, // Precio del producto
-        "idClient": "dory", // ID del cliente (4 letras)
-        "size": "36", // Talle del producto
-        "status": "pendiente", // Estado del producto (pendiente, reservado, cancelado, comprado, recibido)
-        "createdAt": "2025-04-08T15:07:14.976Z", // Fecha de creación del producto
-        "details": "", // Detalles adicionales del producto
-        "color": "negro/rosa" // Color del producto
-    }
-*/
-//--------------------Funciones para la base de datos ubicada en archivo JSON local--------------------
-/*
-import products from '../../jsondata/products.js';
-
-const getAllProducts = () => {
-    return products();
-};
-
-const getProductById = (id) => {
-    console.log(`Buscando producto con ID: ${id}`);
-    const product = products().find((item) => item.id == id);
-    console.log(product);
-    return product;
-};
-
-const postProduct = (newProduct) => {
-    console.log('Nuevo producto recibido:', newProduct);
-    return newProduct; // Simulando que se devuelve el producto creado
-};
-*/
-//----------------Funciones para la base de datos ubicada en base de datos de Firestore-----------
-
 import { db } from '../data/firebase.data.js';
 import {
     collection,
@@ -82,13 +43,11 @@ const getProductById = async (id) => {
     try {
         const snapshot = await getDocs(productsCollection);
         const productDoc = snapshot.docs.find(doc => doc.id == id);
-        if (!productDoc) {
-            return null;
-        }
+        if (!productDoc) { return null }
         const product = { id: productDoc.id, ...productDoc.data() };
         return product;
     } catch (error) {
-        console.error(`Error al buscar producto con id "${id}:`, error);
+        console.error(`Error al buscar producto con id ${id}:`, error);
         throw error;
     }
 };
@@ -96,8 +55,6 @@ const getProductById = async (id) => {
 const createProduct = async (newProduct) => {
     try {
         const newDoc = await addDoc(productsCollection, newProduct);
-        console.log({ id: newDoc.id, ...newProduct });
-        console.log('Mensaje: Producto creado exitosamente');
         return { id: newDoc.id, ...newProduct };
     } catch (error) {
         console.error('Error al crear producto:', error);
@@ -110,12 +67,8 @@ const replaceProduct = async (id, updatedProduct) => {
         const productRef = doc(db, 'products', id);
         const snapshot = await getDoc(productRef);
         const actualStatus = snapshot.data().status;
-        if (actualStatus !== 'pendiente' && actualStatus !== 'cancelado') {
-            console.log('Mensaje: El estado actual del producto no permite el reemplazo del registro')
-            return 403;
-        }
+        if (actualStatus !== 'pendiente' && actualStatus !== 'cancelado') { return 403 }
         await updateDoc(productRef, {...updatedProduct});
-        console.log(`Mensaje: Producto con id '${id}' actualizado exitosamente`);
         return { id, ...updatedProduct };
     } catch (error) {
         console.error(`Error al actualizar producto con id "${id}":`, error);
@@ -127,7 +80,6 @@ const updateProduct = async (id, updatedFields) => {
     try {
         const productRef = doc(db, 'products', id);
         await updateDoc(productRef, updatedFields);
-        console.log(`Mensaje: Producto con id "${id}" actualizado exitosamente`);
         return { id, ...updatedFields };
     } catch (error) {
         console.error(`Mensaje: Error al actualizar producto con id "${id}":`, error);
@@ -138,10 +90,13 @@ const updateProduct = async (id, updatedFields) => {
 const deleteProduct = async (id) => {
     try {
         const productRef = doc(db, 'products', id);
+        const snapshot = await getDoc(productRef);
+        const actualStatus = snapshot.data().status;
+        if (actualStatus == 'comprado' || actualStatus == 'recibido') { return 403 }
         await deleteDoc(productRef);
-        console.log(`Mensaje: Producto con id '${id}' eliminado exitosamente`);
+        return 200;
     } catch (error) {
-        console.error(`Error al eliminar producto con id "${id}":`, error);
+        console.error(`Error al eliminar producto con id ${id}:`, error);
         throw error;
     }
 };
